py/pp-formulanet/cuda/pyproject.toml
[project]
name = "pp-formulanet-cuda"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "pp-formulanet",
    "onnxruntime-gpu>=1.24.2",
    "onnx>=1.20.1",
    "pillow>=12.1.1",
    "tokenizers>=0.21.0",
]

[project.optional-dependencies]
export = ["torch>=2.6.0"]

[tool.uv.sources]
pp-formulanet = { path = "..", editable = true }
